requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
colorama==0.4.6
decorator==5.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
hupper==1.12.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
moviepy==2.1.1
numpy==2.1.3
pillow==10.4.0
proglog==0.1.10
python-dotenv==1.0.1
python-telegram-bot==21.7
setuptools==75.6.0
sniffio==1.3.1
tqdm==4.67.1